# sigpcb-base-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Base",
        "description": "Gestion des API de base",
        "contact": {
            "name": "Support client",
            "email": "contact@exemple.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/anatt-base/arrondissements": {
            "get": {
                "tags": [
                    "Arrondissements"
                ],
                "summary": "Obtient la liste des arrondissements",
                "description": "Obtient la liste des arrondissements enregistrés dans la base de données",
                "operationId": "getArrondissements",
                "responses": {
                    "200": {
                        "description": "Liste des arrondissements",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID de l'arrondissement",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "description": "Nom de l'arrondissement",
                                                "type": "string",
                                                "example": "1er arrondissement"
                                            },
                                            "commune_id": {
                                                "description": "ID de la commune à laquelle appartient l'arrondissement",
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Arrondissements"
                ],
                "summary": "Crée un nouvel arrondissement",
                "description": "Crée un nouvel arrondissement enregistré dans la base de données",
                "operationId": "createArrondissement",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom de l'arrondissement",
                                        "type": "string",
                                        "example": "1er arrondissement"
                                    },
                                    "commune_id": {
                                        "description": "ID de la commune à laquelle appartient l'arrondissement",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nouvel arrondissement créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID du nouvel arrondissement créé",
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "name": {
                                            "description": "Nom du nouvel arrondissement créé",
                                            "type": "string",
                                            "example": "2ème arrondissement"
                                        },
                                        "commune_id": {
                                            "description": "ID de la commune à laquelle appartient le nouvel arrondissement créé",
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/anatt-base/arrondissements/{id}": {
            "get": {
                "tags": [
                    "Arrondissements"
                ],
                "summary": "Affiche les détails d'un arrondissement",
                "description": "Affiche les détails d'un arrondissement enregistré dans la base de données",
                "operationId": "getArrondissementById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'arrondissement à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'arrondissement récupéré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID de l'arrondissement",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "description": "Nom de l'arrondissement",
                                            "type": "string",
                                            "example": "Arrondissement 1"
                                        },
                                        "commune_id": {
                                            "description": "ID de la commune à laquelle appartient l'arrondissement",
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Arrondissement non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Arrondissements"
                ],
                "summary": "Mettre à jour un arrondissement",
                "description": "Met à jour un arrondissement enregistré dans la base de données",
                "operationId": "updateArrondissement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'arrondissement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom de l'arrondissement",
                                        "type": "string",
                                        "example": "Villejuif"
                                    },
                                    "commune_id": {
                                        "description": "ID de la commune associée à l'arrondissement",
                                        "type": "integer",
                                        "example": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Arrondissement mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID de l'arrondissement mis à jour",
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "name": {
                                            "description": "Nom de l'arrondissement mis à jour",
                                            "type": "string",
                                            "example": "Villejuif"
                                        },
                                        "commune_id": {
                                            "description": "ID de la commune associée à l'arrondissement mis à jour",
                                            "type": "integer",
                                            "example": 12
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Arrondissement non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Arrondissements"
                ],
                "summary": "Supprime un arrondissement existant",
                "description": "Supprime l'arrondissement spécifié de la base de données",
                "operationId": "deleteArrondissement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'arrondissement à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arrondissement supprimé avec succès"
                    },
                    "404": {
                        "description": "Arrondissement non trouvé"
                    }
                }
            }
        },
...
    }
}